Make --version a command instead of a class_option
Having --version be a class_option unfortunately results in Thor listing it as a flag for every other command. It appears the best option is to have -v, --version be a command rather than class_option. (An aside if this is made a command, but I made a mistake in the class_option implementation and called `super` on help when it should be `super(*args)`.)